Projeto Conversando com CHAT GPT/requirements.txt
openai
pyaudio
wave
gtts
python-dotenv
pygame